+++ src/NodeDev.Blazor/Services/AppOptions.cs
@@ -1,0 +1,6 @@
+﻿namespace NodeDev.Blazor.Services;
+
+public record class AppOptions
+{
+	public string ProjectsDirectory { get; set; } = string.Empty;
+}
