+++ src/NodeDev.Core/ProjectSettings.cs
@@ -1,7 +1,7 @@
 ﻿namespace NodeDev.Core;
 
-public class ProjectSettings()
+public record class ProjectSettings()
 {
-    public string ProjectName { get; set; } = string.Empty;
-    public static ProjectSettings Default { get; } = new();
+	public string ProjectName { get; set; } = string.Empty;
+	public static ProjectSettings Default { get; } = new();
 }
