+++ src/NodeDev.Tests/ProjectServiceTest.cs
@@ -1,38 +1,32 @@
-﻿using NodeDev.Core;
-using NodeDev.Blazor.Services;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using NodeDev.Core.Class;
+﻿using NodeDev.Blazor.Services;
+using NodeDev.Core;
 
 namespace NodeDev.Tests;
 
 public class ProjectServiceTest
 {
-    [Fact]
-    public void TestsChangeProject()
-    {
-        var project = new Project(Guid.NewGuid());
-        var projectService = new ProjectService();
+	[Fact]
+	public void TestsChangeProject()
+	{
+		var project = new Project(Guid.NewGuid());
+		var projectService = new ProjectService(new AppOptionsContainer(""));
 
-        projectService.ChangeProject(project);
+		projectService.ChangeProject(project);
 
-        Assert.Equal(project, projectService.Project);
-    }
+		Assert.Equal(project, projectService.Project);
+	}
 
-    [Fact]
-    public void TestsProjectChangedEvent()
-    {
-        var project = new Project(Guid.NewGuid());
-        var projectService = new ProjectService();
-        bool isEventTriggered = false;
+	[Fact]
+	public void TestsProjectChangedEvent()
+	{
+		var project = new Project(Guid.NewGuid());
+		var projectService = new ProjectService(new AppOptionsContainer(""));
+		bool isEventTriggered = false;
 
-        projectService.ProjectChanged += () => isEventTriggered = true;
-        projectService.ChangeProject(project);
+		projectService.ProjectChanged += () => isEventTriggered = true;
+		projectService.ChangeProject(project);
 
-        Assert.True(isEventTriggered);
-    }
+		Assert.True(isEventTriggered);
+	}
 }
